% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsInt.R
\name{dsCalculateDensityValues}
\alias{dsCalculateDensityValues}
\title{Calculate density values for data source}
\usage{
dsCalculateDensityValues(dataSourceFileName, nNearestNeighbors)
}
\arguments{
\item{dataSourceFileName}{Name of data source file name}

\item{nNearestNeighbors}{number of used nearest neighbors}
}
\value{
None
}
\description{
Read a data source from a file, calculate density values and write the data source with density values to original file.
Calculated density values are used to evaluate a data source.
}
\examples{
\dontrun{
dsCalculateDensityValues("ds.bin")}
}
