% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsDeactivateColumns}
\alias{dsDeactivateColumns}
\title{Deactivate columns}
\usage{
dsDeactivateColumns(columnVector)
}
\arguments{
\item{columnVector}{Vector of column indices}
}
\value{
None
}
\description{
Deactivate columns in a data source in order to exclude them from training
of generative models. Note that the training function in the package
supports only columns of type R-class numeric, R-type double. All columns of
other type have to be deactivated. The training function in the software
service for accelerated training of generative models supports columns of
any type.
}
\examples{
\donttest{
dsCreateWithDataFrame(iris)
dsDeactivateColumns(c(5))
dsGetInactiveColumnNames()}
}
