% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsDensityValueInverseQuantile}
\alias{dsDensityValueInverseQuantile}
\title{Calculate inverse density value quantile}
\usage{
dsDensityValueInverseQuantile(densityValue)
}
\arguments{
\item{densityValue}{Normalized density value}
}
\value{
Percent value
}
\description{
Calculate inverse density value quantile for a density value.
}
\examples{
\dontrun{
dsRead("ds.bin")
dsDensityValueInverseQuantile(0.5)}
}
