% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsGetActiveColumnNames}
\alias{dsGetActiveColumnNames}
\title{Get active column names}
\usage{
dsGetActiveColumnNames()
}
\value{
Names of active columns
}
\description{
Get active column names of the data source
}
\examples{
\donttest{dsCreateWithDataFrame(inDataFrame = iris) 
dsGetActiveColumnNames()}
}
