% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsGetInactiveColumnNames}
\alias{dsGetInactiveColumnNames}
\title{Get inactive column names}
\usage{
dsGetInactiveColumnNames()
}
\value{
Names of inactive columns
}
\description{
Get inactive column names of the data source
}
\examples{
\donttest{dsCreateWithDataFrame(inDataFrame = iris)
dsGetInactiveColumnNames()}
}
