% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsGetNumberOfRows}
\alias{dsGetNumberOfRows}
\title{Get number of rows}
\usage{
dsGetNumberOfRows()
}
\value{
Number of rows
}
\description{
Get number of rows in a data source
}
\examples{
\donttest{dsCreateWithDataFrame(iris)
dsGetNumberOfRows()}
}
