% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsGetRow}
\alias{dsGetRow}
\title{Get a row in a data source}
\usage{
dsGetRow(index)
}
\arguments{
\item{index}{Index of row}
}
\value{
Row in data source
}
\description{
Get a row in a data source containing values of columns for a row index
}
\examples{
\donttest{dsCreateWithDataFrame(inDataFrame = iris)
dsGetRow(1)}
}
