% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsRead}
\alias{dsRead}
\title{Read a data source from file}
\usage{
dsRead(inFileName)
}
\arguments{
\item{inFileName}{Name of data source file}
}
\value{
None
}
\description{
Read a data source from a file in binary format
}
\examples{
\donttest{dsCreateWithDataFrame(dataFrame = iris)
dsDeactivateColumns(c(5))
dsWrite("iris4d.bin")
dsRead("iris4d.bin")}
}
