% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdCalculateDensityValue}
\alias{gdCalculateDensityValue}
\title{Calculate density value for a data record}
\usage{
gdCalculateDensityValue(dataRecord)
}
\arguments{
\item{dataRecord}{List containing a data record}
}
\value{
Normalized density value number
}
\description{
Calculate density value for a data record.
}
\examples{
\donttest{gdRead("gd.bin")
gdCalculateDensityValue(List(6.1, 2.6, 5.6, 1.4))}
}
