% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdComplete}
\alias{gdComplete}
\title{Complete incomplete data record}
\usage{
gdComplete(dataRecord, useSearchTree = FALSE)
}
\arguments{
\item{dataRecord}{List containing incomplete data record}

\item{useSearchTree}{Boolean value indicating if a search tree should be used.}
}
\value{
List containing completed data record
}
\description{
Search for first nearest neighbor in generative data for incomplete data record containing NA values.
Found row in generative data is then used to replace NA values in inccomplete data record. This function calls
gdKNearestNeighbor() with parameter k equal to 1.
}
\examples{
\dontrun{gdRead("gd.bin")
gdComplete(list(5.1, 3.5, 1.4, NA))}
}
