% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdDensityValueInverseQuantile}
\alias{gdDensityValueInverseQuantile}
\title{Calculate inverse density value quantile}
\usage{
gdDensityValueInverseQuantile(densityValue)
}
\arguments{
\item{densityValue}{Normalized density value}
}
\value{
Percent value
}
\description{
Calculate inverse density value quantile for a density value.
}
\examples{
\dontrun{
gdRead("gd.bin")
gdDensityValueInverseQuantile(0.5)}
}
