% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdDensityValueQuantile}
\alias{gdDensityValueQuantile}
\title{Calculate density value quantile}
\usage{
gdDensityValueQuantile(percent)
}
\arguments{
\item{percent}{Percent value}
}
\value{
Normalized density value quantile number
}
\description{
Calculate density value quantile for a percent value.
}
\examples{
\dontrun{
gdRead("gd.bin")
gdDensityValueQuantile(50)}
}
