% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdGetNumberOfRows}
\alias{gdGetNumberOfRows}
\title{Get number of rows}
\usage{
gdGetNumberOfRows()
}
\value{
Number of rows
}
\description{
Get number of rows in generative data
}
\examples{
\donttest{gdRead("gd.bin")
gdGetNumberOfRows()}
}
