% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdGetRow}
\alias{gdGetRow}
\title{Get a row in generative data}
\usage{
gdGetRow(index)
}
\arguments{
\item{index}{Index of row}
}
\value{
List containing row in generative data
}
\description{
Get a row in generative data for a row index
}
\examples{
\donttest{gdRead("gd.bin")
gdGetRow(1000)}
}
