% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdPlot2dProjection.R
\name{gdPlotDataSourceParameters}
\alias{gdPlotDataSourceParameters}
\title{Specify plot parameters for data source}
\usage{
gdPlotDataSourceParameters(numberOfRandomPoints = 0, color = "blue")
}
\arguments{
\item{numberOfRandomPoints}{Number of randomly selected rows in data source}

\item{color}{Colour for data points of data source}
}
\description{
Specify plot parameters for data source passed to function gdPlot2dProjection().
}
\examples{
\dontrun{gdPlotDataSourceParameters(2500)}
}
