% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdPlotProjection.R
\name{gdRead}
\alias{gdRead}
\title{Read generative data and data source}
\usage{
gdRead(generativeDataFileName, dataSourceFileName = "")
}
\arguments{
\item{generativeDataFileName}{Name of generative data file}

\item{dataSourceFileName}{Name of data source file}
}
\value{
None
}
\description{
Read generative data and data source from specified files. Read in generative data and data source are accessed in gdPlotProjection(),
generative data is accessed in gdGetRow(), gdGetNumberOfRows(), gdCalculateDensityValue(),
gdCalculateDensityValueQuantile(), gdGetNumberOfRows(), gdKNearestNeighbors() and gdComplete().
}
\examples{
\dontrun{
gdRead("gd.bin", "ds.bin")}
}
