% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdService.R
\name{gdServiceDelete}
\alias{gdServiceDelete}
\title{Delete a generated job from software service for accelerated training of
generative models}
\usage{
gdServiceDelete(url, accessKey, jobId)
}
\arguments{
\item{url}{URL of software service for accelerated training of generative
models}

\item{accessKey}{Unique key for access to software service}

\item{jobId}{Job id for generated job for training a generative model}
}
\value{
None
}
\description{
Delete a generated job from software service. If the job is currently
executed it it will be stopped.
}
\examples{
\dontrun{
url <- "http://xxx.xxx.xxx.xxx/gdService"
accessKey <- "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
gdServiceDelete(url, accessKey, 1)}
}
