% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdService.R
\name{gdServiceGetStatus}
\alias{gdServiceGetStatus}
\title{Get status of generated job from software service for accelerated training of
generative models}
\usage{
gdServiceGetStatus(url, accessKey, jobId)
}
\arguments{
\item{url}{URL of software service for accelerated training of generative
models}

\item{accessKey}{Unique key for access to software service}

\item{jobId}{Job id for generated job for training a generative model}
}
\value{
List containing status information
}
\description{
Get status of generated job from software service. Defined status values are:
CREATED, UPLOADED, TRAINING, TRAINED, DELETED, ERROR.
}
\examples{
\dontrun{
url <- "http://xxx.xxx.xxx.xxx/gdService"
accessKey <- "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
gdServiceGetStatus(url, accessKey, 1)}
}
