% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gdWriteSubset}
\alias{gdWriteSubset}
\title{Write subset of generative data}
\usage{
gdWriteSubset(fileName, percent)
}
\arguments{
\item{fileName}{Name of subset generative data file}

\item{percent}{Percent of randomly selected rows}
}
\value{
None
}
\description{
Write subset of randomly selected rows of generative data
}
\examples{
\dontrun{
gdRead("gd.bin")
gdWriteSubset("gds.bin", 50)}
}
