% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsActivateColumns}
\alias{dsActivateColumns}
\title{Activate columns}
\usage{
dsActivateColumns(columnVector)
}
\arguments{
\item{columnVector}{Vector of column indices}
}
\value{
None
}
\description{
Activate columns in a data source in order to include them in training of generative models. By default columns are active.
}
\examples{
\donttest{
dsCreateWithDataFrame(iris)
dsGetActiveColumnNames()
dsDeactivateColumns(c(5))
dsGetActiveColumnNames()
dsActivateColumns(c(5))
dsGetActiveColumnNames()}
}
