% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsInt.R
\name{dsCreateWithDataFrame}
\alias{dsCreateWithDataFrame}
\title{Create a data source with passed data frame}
\usage{
dsCreateWithDataFrame(dataFrame)
}
\arguments{
\item{dataFrame}{Name of data frame}
}
\value{
None
}
\description{
Create a data source with passed data frame.
}
\examples{
\donttest{
dsCreateWithDataFrame(iris)}
}
