% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsDeactivateColumns}
\alias{dsDeactivateColumns}
\title{Deactivate columns}
\usage{
dsDeactivateColumns(columnVector)
}
\arguments{
\item{columnVector}{Vector of column indices}
}
\value{
None
}
\description{
Deactivate columns of the data source in order to exclude them in generation of generative data.
Note that in this version only columns with values of type double or float can be used in generation of generative data.
All columns with values of other type have to be deactivated.
}
\examples{
\donttest{dsCreateWithDataFrame(inDataFrame = iris)
dsDeactivateColumns(c(5))
dsGetInactiveColumnNames()}
}
