% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsGetActiveColumnNames}
\alias{dsGetActiveColumnNames}
\title{Get active column names}
\usage{
dsGetActiveColumnNames()
}
\value{
Vector of names of active columns
}
\description{
Get active column names of a data source
}
\examples{
\donttest{dsCreateWithDataFrame(dataFrame = iris)
dsDeactivateColumns(c(5))
dsGetActiveColumnNames()}
}
