% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsGetRow}
\alias{dsGetRow}
\title{Get a row in a data source}
\usage{
dsGetRow(index)
}
\arguments{
\item{index}{Index of row}
}
\value{
List containing row in data source
}
\description{
Get a row in a data source for a row index.
}
\examples{
\donttest{dsCreateWithDataFrame(dataFrame = iris)
dsGetRow(1000)}
}
