% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsWrite}
\alias{dsWrite}
\title{Write a data source to file}
\usage{
dsWrite(outFileName)
}
\arguments{
\item{outFileName}{Name of data source file}
}
\value{
None
}
\description{
Write a data source to a file in binary format
}
\examples{
\donttest{dsCreateWithDataFrame(inDataFrame = iris)
dsDeactivateColumns(c(5))
dsWrite("iris4d.bin")}
}
