% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdPlotProjection.R
\name{gdPlotDataSourceParameters}
\alias{gdPlotDataSourceParameters}
\title{Specify plot parameters for data source}
\usage{
gdPlotDataSourceParameters(percent = 100, color = "blue")
}
\arguments{
\item{percent}{Percent of randomly selected rows in data source}

\item{color}{Colour for data points of data source}
}
\value{
List of plot parameters for data source
}
\description{
Specify plot parameters for data source passed to function gdPlotProjection().
}
\examples{
\dontrun{
gdPlotDataSourceParameters(2500)}
}
