% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdPlot2dProjection.R
\name{gdRead}
\alias{gdRead}
\title{Read generative data and data source}
\usage{
gdRead(inGenerativeDataFileName, inDataSourceFileName = "")
}
\arguments{
\item{inGenerativeDataFileName}{Name of generative data file}

\item{inDataSourceFileName}{Name of data source file}
}
\value{
None
}
\description{
Read generative data and data source from corresponding files. Read in generative data and data source are accessed in gdPlot2dProjection(),
generative data in gdGetRow().
}
\examples{
\donttest{gdRead("gd.bin", "iris4d.bin")}
}
