% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdPlot2dProjection.R
\name{gdRead}
\alias{gdRead}
\title{Read generative data and data source}
\usage{
gdRead(generativeDataFileName, dataSourceFileName = "")
}
\arguments{
\item{generativeDataFileName}{Name of generative data file}

\item{dataSourceFileName}{Name of data source file}
}
\value{
None
}
\description{
Read generative data and data source from specified files. Read in generative data and data source are accessed in gdPlot2dProjection(),
generative data is accessed in gdGetRow(), gdCalculateDensityValue() and gdCalculateDensityValueQuantile().
}
\examples{
\dontrun{gdRead("gd.bin", "iris4d.bin")}
}
